requirements.txt
django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
pymysql==1.1.1
python-dotenv==1.1.0
openai>=1.0.0
pypdf>=3.0.0
requests>=2.31.0
wechatpayv3>=1.2.0
cryptography>=41.0.0
celery>=5.3.0
redis>=5.0.0
django-celery-results>=2.5.0
google-generativeai>=0.3.0
youtube-transcript-api>=0.6.0
azure-cognitiveservices-speech>=1.31.0
dashscope>=1.25.0
google-genai
Pillow
PyMuPDF>=1.26.0
